% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{set_api_key}
\alias{set_api_key}
\title{Set Google Places API (New) key}
\usage{
set_api_key(key = NULL)
}
\arguments{
\item{key}{See \href{https://developers.google.com/maps/documentation/places/web-service/get-api-key}{here} how to get an api key from google.}
}
\description{
Set Google Places API (New) key
}
